saturd : create option

+++ include/print-reply.h
@@ -7,7 +7,7 @@
 
 void print_reply_l(uint32_t, task**);
 
-void print_reply_c(uint32_t);
+void print_reply_c(uint64_t);
 
 void print_error(uint16_t);
